add svelte useKeymap
