+++ packages/svelte/src/contexts/editor-context.ts
@@ -1,11 +1,8 @@
 import {
   ProseKitError,
   type Editor,
-  type Extension,
-  defineUpdateHandler,
 } from '@prosekit/core'
-import { getContext, hasContext, onDestroy, onMount, setContext } from 'svelte'
-import { readonly, writable, type Readable } from 'svelte/store'
+import { getContext, hasContext, setContext } from 'svelte'
 
 export interface EditorContext {
   editor: Editor
@@ -30,23 +27,4 @@ export function getEditorContext(): EditorContext {
   return getContext(key)
 }
 
-export function getEditor<E extends Extension = any>(options?: {
-  update?: boolean
-}): Readable<Editor<E>> {
-  const update = options?.update ?? false
-  const editor = getEditorContext().editor as Editor<E>
 
-  const editorStore = writable(editor)
-
-  if (update) {
-    onMount(() => {
-      const forceUpdate = () => {
-        editorStore.set(editor)
-      }
-      const dispose = editor.use(defineUpdateHandler(forceUpdate))
-      onDestroy(dispose)
-    })
-  }
-
-  return readonly(editorStore)
-}
