+++ packages/svelte/src/hooks/use-extension.ts
@@ -1,0 +1,16 @@
+import { type Extension } from '@prosekit/core'
+import type { Readable } from 'svelte/store'
+
+import { useEditor } from './use-editor'
+
+export function useExtension<T extends Extension = Extension>(
+  extensionStore: Readable<T>,
+) {
+  const editorStore = useEditor()
+
+  return extensionStore.subscribe((extension) => {
+    return editorStore.subscribe((editor) => {
+      return editor.use(extension)
+    })
+  })
+}
